+++ src/namespaces/emergency.ts
@@ -2,7 +2,7 @@ import { getRandomString, Header, } from '../utils';
 import type { PidfLo } from 'pidf-lo';
 import PidfLoCompat from '../compatibility/pidf-lo';
 import { PIDF_LO, TEXT_PLAIN, CALL_SUB, TEXT_URI_LIST, MULTIPART_MIXED } from '../constants/content-types';
-import { CALL_INFO, CONTENT_ID, CONTENT_TYPE, GEOLOCATION, GEOLOCATION_ROUTING, REPLY_TO } from '../constants/headers';
+import { CALL_INFO, CONTENT_ID, CONTENT_TYPE, GEOLOCATION, GEOLOCATION_ROUTING, HISTORY_INFO, REPLY_TO } from '../constants/headers';
 import { ConversationEndpointType } from '../models/conversation';
 import { CRLF, Multipart, MultipartPart } from '../models/multipart';
 import { VCard, VCARD_XML_NAMESPACE } from '../models/vcard';
@@ -65,6 +65,7 @@ export class EmergencyMapper implements NamespacedConversation {
   getIsTestFromEvent = (evt: NewMessageEvent): boolean => false; // TODO: reference ETSI TS 103 698, 6.1.2.10 "Test Call"
 
   protected createCommonParts = (
+    targetUri: string,
     endpointType: ConversationEndpointType,
     replyToSipUri: string,
     text?: string,
@@ -72,10 +73,13 @@ export class EmergencyMapper implements NamespacedConversation {
     extraParts?: MultipartPart[],
     location?: PidfLo,
     vcard?: VCard,
-    // TODO: TS 103 698 -> 6.1.2.11 -> support DIDs
   ): MessageParts => {
 
-    let headers: Header[] = [];
+    let headers: Header[] = [
+      // enables tracing back the origin and routing history of the call
+      // according to ETSI TS 103 698 -> 6.1.2.6
+      { key: HISTORY_INFO, value: `<${targetUri}>;index=1` }
+    ];
     const multipart = new Multipart();
 
     if (extraParts) {
@@ -259,6 +263,7 @@ export class EmergencyMapper implements NamespacedConversation {
   }
 
   createMessageParts = ({
+    targetUri,
     conversationId,
     isTest,
     id,
@@ -273,6 +278,7 @@ export class EmergencyMapper implements NamespacedConversation {
     did,
   }: MessagePartsParams): MessageParts => {
     const parts = this.createCommonParts(
+      targetUri,
       endpointType,
       replyToSipUri,
       text,